# graph/model/models_gen.go
// Code generated by github.com/99designs/gqlgen, DO NOT EDIT.

package model

type Country struct {
	Name string `json:"name"`
}

type CountryInput struct {
	UserID int    `json:"userId"`
	Name   string `json:"name"`
}

type LoginInput struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

type PercentageInput struct {
	UserID int    `json:"userId"`
	Name   string `json:"name"`
}

type RegisterInput struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

type TopThreeCountriesInput struct {
	UserID int    `json:"userId"`
	Type   string `json:"type"`
}

type User struct {
	ID       string `json:"id"`
	Email    string `json:"email"`
	Password string `json:"password"`
}